add comment

+++ main_bonus.c
@@ -6,7 +6,7 @@
 /*   By: hyowchoi <hyowchoi@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/13 12:28:07 by hyowchoi          #+#    #+#             */
-/*   Updated: 2023/12/26 16:00:26 by hyowchoi         ###   ########.fr       */
+/*   Updated: 2023/12/27 12:34:53 by hyowchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,6 +18,7 @@ static int	fin_child_process(t_defaults def, int end, int last_child_pid)
 	int	status;
 	int	value;
 
+	// if here_doc, child_cnt is 1 small (argc cnt is 1 large)
 	if (ft_strcmp(def.argv[1], "here_doc") == 0)
 		end--;
 	i = 0;
@@ -40,6 +41,7 @@ int	main(int argc, char **argv, char **env)
 	if (argc < 5)
 		print_error_n_exit(ARGUMENT_ERROR);
 	make_struct_n_pipe(&def, argc, argv, env);
+
 	if (ft_strcmp(def.argv[1], "here_doc") == 0)
 	{
 		file_name = make_file_name("heredoc_tmp_file");
@@ -47,8 +49,14 @@ int	main(int argc, char **argv, char **env)
 	}
 	else
 		last_child_pid = do_child(def, argv[1], 2);
+
 	close_pipes(def.pipes, def.cnt_pipes, 0);
+
+	// find last_child's return code
 	value = fin_child_process(def, argc - 3, last_child_pid);
-	unlink(file_name);
+
+	// remove tmp file
+	if (ft_strcmp(def.argv[1], "here_doc") == 0)
+		unlink(file_name);
 	exit (value);
 }
